Split code into utils file
