+++ nvsvocprez/view/templates/concept.html
@@ -112,7 +112,8 @@
               </tr>
             {% else %}
               <tr class="group-header header">
-                <th colspan="3" style="color:#007dbb;font-size:12px;">
+                <th></th>
+                <th colspan="3" style="color:#007dbb;font-size:12px;cursor: pointer">
                   {{ "{} {} - ({})".format(k, alt_labels[k], vi|length) }}
                   <span class="toggle-icon">[+]</span>
                 </th>
